refactor extending vega lite
